Fixed the result from market cap where the capitalization is returned as double instead of integer.

+++ lib/pages/insight/insight_broker.dart
@@ -531,7 +531,7 @@ class _InsightBrokerPageState extends State<InsightBrokerPage> {
                           no: (index + 1),
                           code: _marketCap[index].code,
                           price: _marketCap[index].lastPrice,
-                          marketCap: _marketCap[index].capitalization,
+                          marketCap: _marketCap[index].capitalization.toInt(),
                           shareOut: _marketCap[index].shareOut
                         );
                       }),
